# Profile: Faster data dict lookup